Fix Units::get_moves()
Since this is a bugfix, I've replicated this change on this branch as well. I assume this will probably lead to a merge conflict when the changes on dev are ready, since the corresponding file in dev was moved, which it appears registered as deleting it and re-adding it - in which case, use those ones and don't worry about this commit.